+++ src/pages/Snapshot.jsx
@@ -1,4 +1,5 @@
 import React, { useEffect, useState } from 'react';
+import { useParams, Link as RouterLink } from 'react-router-dom';
 import { RiAddFill } from 'react-icons/ri';
 import { Flex, Stack, Button, Link, Spinner } from '@chakra-ui/react';
 
@@ -10,6 +11,7 @@ import TextBox from '../components/TextBox';
 import { getSnapshotProposals } from '../utils/theGraph';
 
 const Snapshot = ({ isMember, daoMetaData }) => {
+  const { daoid, daochain } = useParams();
   const [loading, setLoading] = useState(true);
   const [snapshots, setSnapshots] = useState({});
   const { errorToast } = useOverlay();
@@ -43,14 +45,24 @@ const Snapshot = ({ isMember, daoMetaData }) => {
   }, [daoMetaData?.boosts]);
 
   const newSnapshotButton = isMember && (
-    <Button
-      as={Link}
-      href={`https://snapshot.org/#/${daoMetaData?.boosts?.SNAPSHOT?.metadata?.space}/create`}
-      rightIcon={<RiAddFill />}
-      isExternal
-    >
-      New Snapshot
-    </Button>
+    <Flex>
+      <Button
+        as={Link}
+        href={`https://snapshot.org/#/${daoMetaData?.boosts?.SNAPSHOT?.metadata?.space}/create`}
+        rightIcon={<RiAddFill />}
+        isExternal
+        mr={10}
+      >
+        New Snapshot
+      </Button>
+
+      <Button
+        as={RouterLink}
+        to={`/dao/${daochain}/${daoid}/boost/snapshot/settings`}
+      >
+        Boost Settings
+      </Button>
+    </Flex>
   );
 
   return (
